@@ -1,4 +1,5 @@
-import { HEROES } from './../mock-heroes';
+import { MessageService } from './../message.service';
+import { HeroService } from './../hero.service';
 import { Component, OnInit } from '@angular/core';
 import { Hero } from '../hero';
 
@@ -8,14 +9,14 @@ import { Hero } from '../hero';
   styleUrls: ['./heroes.component.css']
 })
 export class HeroesComponent implements OnInit {
-  heroes = HEROES;
-  selectedHero: Hero;
+  heroes: Hero[];
 
-  onSelect(hero: Hero): void {
-      this.selectedHero = hero;
+  getHeroes(): void {
+    this.heroService.getHeroes()
+        .subscribe(heroes => this.heroes = heroes);
   }
 
-  constructor() { }
-  ngOnInit() { }
+  constructor(private heroService: HeroService) { }
+  ngOnInit() {this.getHeroes(); }
 
 }
